LayoutTests/inspector/debugger/source-frame-count-expected.txt

Tests that scripts panel does not create too many source frames.

Debugger was enabled.

Running: testSourceFramesCount
Page reloaded.
Debugger was disabled.
